Changing strucutre. Creating category module, implementing feign on expense-management to consume categogory-management services.
