+++ expense-management/.gitignore
@@ -1,0 +1,4 @@
+/target/
+*.classpath
+*.project
+.settings/
