+++ expense-management/src/main/resources/application.properties
@@ -1,0 +1,21 @@
+server.port=8081
+#mysql config
+spring.jpa.hibernate.ddl-auto=update
+spring.datasource.driverClassName=com.mysql.jdbc.Driver
+spring.jpa.database-platform=org.hibernate.dialect.MySQL5InnoDBDialect
+spring.datasource.url=jdbc:mysql://localhost:3306/db_example?useLegacyDatetimeCode=false&serverTimezone=UTC
+spring.datasource.username=root
+spring.datasource.password=password
+
+
+#Pools 
+spring.datasource.tomcat.max-active=100000
+spring.datasource.tomcat.max-wait=10000
+
+#redis
+spring.redis.host=localhost
+spring.redis.port=6379
+
+#Spring
+spring.application.name=expense-management
+
