--- src/main/java/br/com/camaroti/alex/res/api/controller/CategoryController.java
@@ -1,26 +1,0 @@
-package br.com.camaroti.alex.res.api.controller;
-
-import java.io.IOException;
-import java.util.List;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.ResponseBody;
-import org.springframework.web.bind.annotation.RestController;
-
-import br.com.camaroti.alex.res.api.domain.Category;
-import br.com.camaroti.alex.res.api.service.CategoryService;
-
-@RestController
-public class CategoryController {
-	
-	@Autowired
-	private CategoryService categoryService;
-	
-	@GetMapping(path="/categories/suggest")
-	public @ResponseBody List<Category> suggestCategory(@RequestParam(value="name", required = false, defaultValue = "") String name) throws IOException {
-		return categoryService.findByNameContaining(name);
-	}
-
-}
