--- src/main/java/br/com/camaroti/alex/res/api/repository/CategoryRepository.java
@@ -1,13 +1,0 @@
-package br.com.camaroti.alex.res.api.repository;
-
-import java.util.List;
-
-import org.springframework.data.repository.CrudRepository;
-
-import br.com.camaroti.alex.res.api.domain.Category;
-
-public interface CategoryRepository extends CrudRepository<Category, Integer> {
-
-	Category findByNameIgnoreCase(String name);
-	List<Category> findByNameContainingIgnoreCase(String name);	
-}
